Moved some methods to Utils class for Wrangler.  Made `gulpw config` callable without a gulpw config.
