+++ docs/setup.md
@@ -1,7 +1,6 @@
 ### Setup
 1. Create your project's bundle configs folder;  E.g., './gulpw-bundle-configs' etc.
-2. Create empty `gulpw-config.*` config file at your projects root and populate it either manually or by running
- `gulpw config` (interactive setup for the gulpw-config.* file).
+2. Create empty `gulpw-config.*` config file or run `gulpw config` in your projects root to generate one.
 3. Tell your `gulpw-config.*` file where your bundle configs folder is (set `bundlesPath` to your bundles config path inside config file).
 5. Configure your global tasks within your `gulpw-config.*` file.
 6. (Optional) Execute `gulpw deploy-config` to configure servers to deploy your work to.
