+++ misc/notes.txt
@@ -53,3 +53,5 @@ I.)  Brainstorming:
 #        #  jsp: ...
 #        #  php: ...
 #
+
+Add Task types
