+++ src/WranglerBaseStrategy.js
@@ -1,0 +1,3 @@
+/**
+ * Created by edelacruz on 6/9/2015.
+ */
